Load cache uses array of structs instead of array of pointers to structs
Pgxscan supports parsing results from database into a map, where the
field name is the key and the value is the database value, each map
entry representing a database row. That built in behavior does not
trigger if we pass an array of pointers to structs ([]*mytype), instead
of an array of structs ([]mytype). Therefore. code in loadCache has
been modified to use structs array. This should not make a difference
to the API users, since such array is internal to the function.